Move getMemberProperties to coffeeChatAPI
