+++ backend/src/dao/CoffeeChatDao.ts
@@ -1,6 +1,11 @@
 import { v4 as uuidv4 } from 'uuid';
-import { memberCollection, coffeeChatsCollection, db } from '../firebase';
-import { DBCoffeeChat } from '../types/DataTypes';
+import {
+  memberCollection,
+  coffeeChatsCollection,
+  db,
+  memberPropertiesCollection
+} from '../firebase';
+import { DBCoffeeChat, DBMemberProperties } from '../types/DataTypes';
 import { getMemberFromDocumentReference } from '../utils/memberUtil';
 import BaseDao, { FirestoreFilter } from './BaseDao';
 import { deleteCollection } from '../utils/firebase-utils';
@@ -136,5 +141,17 @@ export default class CoffeeChatDao extends BaseDao<CoffeeChat, DBCoffeeChat> {
    */
   static async getCoffeeChatBingoBoard(): Promise<string[][]> {
     return COFFEE_CHAT_BINGO_BOARD;
+  }
+
+  /**
+   * Gets the properties for a specific member
+   * @param email - the email of the member whose properties we want to retrieve.
+   * @returns A promise that resolves to an DBMemberProperties object or undefined.
+   */
+  static async getMemberProperties(email: string): Promise<DBMemberProperties | undefined> {
+    return memberPropertiesCollection
+      .doc(email)
+      .get()
+      .then((docRef) => docRef.data());
   }
 }
